Strict loading for all models (#249)
* Initial commit for strict loading

* Strict loading for company model

* Strict loading for alternative name model

* Strict loading for cast member model

* Strict loading for certification model

* Strict loading for company assignment model

* Strict loading for confirmation token model

* Strict loading for content rating model

* Strict loading for country model

* Strict loading for crew member model

* Strict loading for episode model

* Strict loading for genre model

* Strict loading for genre assignment model

* Strict loading for job model

* Strict loading for language model

* Strict loading for movie model

* Strict loading for person model

* Strict loading for release model

* Strict loading for season model

* Strict loading for session model

* Strict loading for show model

* Strict loading for tagline model

* Strict loading for user model

* Strict loading for video model

* All models now use strict loading by default

* Revert "All models now use strict loading by default"

This reverts commit 4016b6b45003561a8de8095defe1596b8c12c55b.

* Move strict_loading_by_default into ApplicationRecord

* Do not raise strick loading violations in console

* Include genres in shows controller

* Some more fixes

+++ app/jobs/prune_unconfirmed_users_job.rb
@@ -3,7 +3,7 @@ class PruneUnconfirmedUsersJob < ApplicationJob
 
   def perform(*args)
     ConfirmationToken.stale.destroy_all
-    User.stale.destroy_all
+    User.includes(:confirmation_tokens, :sessions).stale.destroy_all
     User.includes(:confirmation_tokens).needing_confirmation_reminder.each do |user|
       confirmation_token = user.confirmation_tokens.active.first || user.confirmation_tokens.create
       ConfirmationsMailer.reminder(user, confirmation_token).deliver_later
